merge upstream changes

+++ platformio/src/config.cpp
@@ -115,14 +115,24 @@ const char *NTP_SERVER_2 = "time.nist.gov";
 // NTP_TIMEOUT or select closer/lower latency time servers.
 const unsigned long NTP_TIMEOUT = 20000; // ms
 // Sleep duration in minutes. (aka how often esp32 will wake for an update)
-// Aligned to the nearest minute boundary and must evenly divide 60.
+// Aligned to the nearest minute boundary.
 // For example, if set to 30 (minutes) the display will update at 00 or 30
-// minutes past the hour. (range: [2-60])
-const long SLEEP_DURATION = 30;
+// minutes past the hour. (range: [2-1440])
+// Note: The OpenWeatherMap model is update every 10 minutes, so updating more
+//       frequently than that is unnessesary.
+const int SLEEP_DURATION = 30; // minutes
+// Bed Time Power Savings.
 // If BED_TIME == WAKE_TIME, then this battery saving feature will be disabled.
 // (range: [0-23])
 const int BED_TIME  = 00; // Last update at 00:00 (midnight) until WAKE_TIME.
 const int WAKE_TIME = 06; // Hour of first update after BED_TIME, 06:00.
+// Note that the minute alignment of SLEEP_DURATION begins at WAKE_TIME even if
+// Bed Time Power Savings is disabled.
+// For example, if WAKE_TIME = 00 (midnight) and SLEEP_DURATION = 120, then the
+// display will update at 00:00, 02:00, 04:00... until BED_TIME.
+// If you desire to have your display refresh exactly once a day, you should set
+// SLEEP_DURATION = 1440, and you can set the time it should update each day by
+// setting both BED_TIME and WAKE_TIME to the hour you want it to update.
 
 // HOURLY OUTLOOK GRAPH
 // Number of hours to display on the outlook graph. (range: [8-48])
